Refactor gallery landing page to use custom components for gallery views
