+++ src/components/GalleryItem.js
@@ -1,0 +1,74 @@
+import styled from 'styled-components';
+import tinyColor from 'tinycolor2';
+import Link from 'gatsby-link';
+import GatsbyImage from '../components/GatsbyImage';
+import { BODY_WEIGHTS } from '../styles/vars';
+
+const Content = styled.div`
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  z-index: 2;
+  padding: 0.5rem;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  transition: transform 0.3s ease-in;
+  transform: translateY(100%);
+
+  p,
+  h4 {
+    margin-bottom: 0;
+    padding-left: 0.25rem;
+    padding-right: 0.25rem;
+
+    background-color: ${tinyColor('#fff')
+      .setAlpha(0.6)
+      .toRgbString()};
+    text-shadow: 0 0 1px rgba(255, 255, 255, 0.3);
+    font-weight: ${BODY_WEIGHTS.medium};
+  }
+
+  h4 {
+    text-transform: uppercase;
+  }
+`;
+
+const GalleryItem = styled(Link)`
+  position: relative;
+  display: flex;
+  height: 100%;
+  overflow: hidden;
+
+  &:not(:last-child) {
+    margin-bottom: 0;
+  }
+
+  .gatsby-image-outer-wrapper {
+    height: 100%;
+    width: 100%;
+  }
+
+  &:hover,
+  &:focus {
+    ${Content} {
+      transform: translateY(0);
+    }
+  }
+`;
+
+const Image = GatsbyImage.extend`
+  height: 100%;
+  z-index: 1;
+  transition: transform 3s;
+
+  &:hover {
+    transform: scale(1.15);
+  }
+`;
+
+GalleryItem.Image = Image;
+GalleryItem.Content = Content;
+
+export default GalleryItem;
